ECOMPHP-232
Stopping finalizations if orders are frozen on internal getPayments before the real finalization gets through. If finalization are sent to Resurs Bank even if they are frozen, Resurs Bank will throw a similar error. This is useless and finalizations takes longer time to be completed, from a performance point of view.