ECOMPHP-343: Testcase modified. Rescuer rescued.
1) getNextInvoiceNumberByDebits

While setting new invoice number on demand, ECom also fetches what's been set. While in paranoid mode (getNextInvoiceNumberByDebits) we choose to pick this id if its larger than the old $properInvoiceNumber, and then rerunning the getNextInvoiceNumber() with the new id+1. If $properInvoicenumber is larger, $proverInvoicenumber will increment by 1 instead.

2) finalizeWithoutInvoiceId: Code 500 prevention

Code 500 halts the current test if anything goes wrong in the current test. If something is wrong in this test on the API side we don't want to proceed as it is extremely time consuming. Test will be marked as skipped. This includes the four inital bookPayments.

3) Unexpected results

If the $expectedAssertResult fails, the list of assertions will be displayed in the test.